feat: remove publish button from share popover
