+++ packages/ai-workspace-common/src/components/canvas/nodes/memo/memo.tsx
@@ -483,6 +483,7 @@ export const MemoNode = ({ data, selected, id, isPreview = false, onNodeClick }:
                 arrow={false}
                 open={dropdownOpen}
                 onOpenChange={setDropdownOpen}
+                getPopupContainer={() => targetRef.current}
               >
                 <Button
                   type="text"
