+++ packages/ai-workspace-common/src/components/canvas/top-toolbar/toolbar-buttons.tsx
@@ -8,6 +8,7 @@ import { ToolsDependency } from '../tools-dependency';
 import { CreateVariablesModal } from '@refly-packages/ai-workspace-common/components/canvas/workflow-variables';
 import { genMemoID } from '@refly/utils/id';
 import { useAddNode } from '@refly-packages/ai-workspace-common/hooks/canvas/use-add-node';
+import { useCanvasContext } from '@refly-packages/ai-workspace-common/context/canvas';
 
 interface ToolbarButtonsProps {
   canvasId: string;
@@ -16,6 +17,7 @@ interface ToolbarButtonsProps {
 export const ToolbarButtons = memo(({ canvasId }: ToolbarButtonsProps) => {
   const { t } = useTranslation();
   const { addNode } = useAddNode();
+  const { readonly } = useCanvasContext();
 
   const [showCreateVariablesModal, setShowCreateVariablesModal] = useState(false);
 
@@ -65,29 +67,38 @@ export const ToolbarButtons = memo(({ canvasId }: ToolbarButtonsProps) => {
     createMemo({ x: 0, y: 0 });
   }, []);
 
+  const internalActions = useMemo(() => {
+    return readonly
+      ? []
+      : [
+          {
+            key: 'addUserInput',
+            icon: <Chat size={16} />,
+            onClick: handleAddUserInput,
+            label: t('canvas.toolbar.tooltip.addUserInput'),
+            active: false,
+          },
+          {
+            key: 'addAgent',
+            icon: <AiChat size={16} />,
+            onClick: handleAddAgent,
+            label: t('canvas.toolbar.tooltip.addAgent'),
+            active: false,
+          },
+          {
+            key: 'addNote',
+            icon: <Note1 size={16} />,
+            onClick: handleAddMemo,
+            label: t('canvas.toolbar.tooltip.addNote'),
+          },
+          { type: 'divider', key: 'divider-1' },
+        ];
+  }, [handleAddUserInput, handleAddAgent, handleAddMemo, t, readonly]);
+
   const actions = useMemo(() => {
     return [
-      {
-        key: 'addUserInput',
-        icon: <Chat size={16} />,
-        onClick: handleAddUserInput,
-        label: t('canvas.toolbar.tooltip.addUserInput'),
-        active: false,
-      },
-      {
-        key: 'addAgent',
-        icon: <AiChat size={16} />,
-        onClick: handleAddAgent,
-        label: t('canvas.toolbar.tooltip.addAgent'),
-        active: false,
-      },
-      {
-        key: 'addNote',
-        icon: <Note1 size={16} />,
-        onClick: handleAddMemo,
-        label: t('canvas.toolbar.tooltip.addNote'),
-      },
-      { type: 'divider', key: 'divider-1' },
+      ...internalActions,
+
       {
         key: 'resources',
         icon: <Resource size={16} />,
@@ -96,14 +107,7 @@ export const ToolbarButtons = memo(({ canvasId }: ToolbarButtonsProps) => {
         active: sidePanelVisible,
       },
     ];
-  }, [
-    handleAddUserInput,
-    handleAddAgent,
-    handleAddMemo,
-    handleResourcesPanelOpen,
-    t,
-    sidePanelVisible,
-  ]);
+  }, [internalActions, handleResourcesPanelOpen, t, sidePanelVisible]);
 
   return (
     <div className="absolute bottom-6 left-0 right-0 z-20 p-2 flex items-center justify-center pointer-events-none">
@@ -126,15 +130,18 @@ export const ToolbarButtons = memo(({ canvasId }: ToolbarButtonsProps) => {
         <Divider type="vertical" className="m-0 h-5 bg-refly-Card-Border" />
 
         <ToolsDependency canvasId={canvasId} />
+        {readonly && <div />}
 
-        <Button
-          type="primary"
-          icon={<Play size={16} />}
-          onClick={handleShowWorkflowRun}
-          className="px-[10px] py-[5px] font-semibold ml-4"
-        >
-          {t('canvas.toolbar.tooltip.previewWorkflowRun')}
-        </Button>
+        {!readonly && (
+          <Button
+            type="primary"
+            icon={<Play size={16} />}
+            onClick={handleShowWorkflowRun}
+            className="px-[10px] py-[5px] font-semibold ml-4"
+          >
+            {t('canvas.toolbar.tooltip.previewWorkflowRun')}
+          </Button>
+        )}
       </div>
 
       <CreateVariablesModal
